requirements.txt
fastapi
uvicorn
Pillow
pikepdf
python-multipart
brotli
requests
lxml
bsdiff4
python-dotenv
flif
pyheif
slowapi
